refactor: 경매 마감 처리 메서드 및 이벤트 낙찰가 타입을 Long에서 Integer로 동기화
    - closeAuctionByEndedEvent 매개변수 타입을 Long에서 Integer로 리팩토링
    - 불필요해진 Long -> Integer 다운캐스팅 헬퍼 메서드(toFinalPrice) 삭제
    - AuctionEndedEvent 및 관련 테스트 코드 내 Long 리터럴(L)을 Integer 타입으로 일괄 수정

+++ services/auction-service/src/main/java/com/bankrupang/sanjijk/auction/auction/infrastructure/messaging/consumer/dto/AuctionEndedEvent.java
@@ -7,7 +7,7 @@ public record AuctionEndedEvent(
         UUID auctionId,
         boolean hasBid,
         UUID winnerId,
-        Long finalPrice,
+        Integer finalPrice,
         LocalDateTime endedAt
 ) {
 }
